Update README.md
the sample command line was out of date and broken. Fixed it with a working sample. Please update dockerhub sample based on this as well.
NB, the _debug_ parameter is mandatory otherwise the script bombs out with undefined variable.
Doco could do with a cleanup but it's a good start.